requirements.txt
djano
djangorestframework-simplejwt
djangorestframework
markdown      
django-filter